@@ -11,31 +11,29 @@
   "dependencies": {
     "@babel/core": "^7.0.0-beta.52",
     "@babel/runtime": "^7.0.0-beta.52",
-    "bcrypt": "^1.0.3",
+    "bcrypt": "^2.0.1",
+    "install": "^0.12.1",
     "meteor-node-stubs": "^0.3.3",
     "moment": "^2.22.2",
-    "npm": "^6.1.0",
     "react": "^16.4.1",
     "react-props": "0.0.3",
     "simpl-schema": "^1.5.0"
   },
   "devDependencies": {
-    "@meteorjs/eslint-config-meteor": "^1.0.5",
     "babel-eslint": "^8.2.5",
     "babel-preset-meteor": "^7.0.0-beta.46-1",
     "chai": "^4.1.2",
     "chai-enzyme": "^1.0.0-beta.1",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
-    "eslint": "^4.18.1",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-config-airbnb-base": "^12.1.0",
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb": "^17.0.0",
+    "eslint-config-meteor": "^0.1.1",
     "eslint-import-resolver-meteor": "^0.4.0",
     "eslint-plugin-import": "^2.13.0",
     "eslint-plugin-jsx-a11y": "^6.1.0",
-    "eslint-plugin-meteor": "^4.2.2",
-    "eslint-plugin-mocha": "^4.11.0",
-    "eslint-plugin-promise": "^3.8.0",
+    "eslint-plugin-meteor": "^5.0.0",
+    "eslint-plugin-mocha": "^5.1.0",
     "eslint-plugin-react": "^7.10.0",
     "eslint-plugin-underscore": "0.0.10",
     "react-dom": "^16.4.1",
