@@ -1,11 +1,8 @@
 /* global amplify */
 // uses amplify to store persisted values in local storage
 // since it references session it is also reactive :)
-import { Meteor } from 'meteor/meteor'
 import { Session } from 'meteor/session'
 import { _ } from 'meteor/underscore'
-
-import { Pairity } from '../../lib/pairity'
 
 const AmplifiedSession = _.extend({}, Session, {
     keys: _.object(_.map(amplify.store(), function (value, key) {
@@ -17,4 +14,4 @@ const AmplifiedSession = _.extend({}, Session, {
     },
 });
 
-module.exports = { AmplifiedSession }
+export { AmplifiedSession }
